+++ Patches/HealthUIPatch.cs
@@ -1,11 +1,4 @@
-﻿using ExitGames.Client.Photon;
-using HarmonyLib;
-using MenuLib;
-using REPOLib.Modules;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using UnityEngine;
+﻿using HarmonyLib;
 
 namespace MapVote.Patches
 {