+++ Patches/SteamManagerPatch.cs
@@ -1,10 +1,6 @@
 ﻿using ExitGames.Client.Photon;
 using HarmonyLib;
 using REPOLib.Modules;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using UnityEngine;
 
 namespace MapVote.Patches
 {
@@ -18,6 +14,7 @@ namespace MapVote.Patches
             if (SemiFunc.IsMasterClient())
             {
                 MapVote.OnSyncVotes?.RaiseEvent(MapVote.CurrentVotes.Values, NetworkingEvents.RaiseAll, SendOptions.SendReliable);
+                MapVote.OnSyncLastMapPlayed?.RaiseEvent(MapVote.LastMapPlayed, NetworkingEvents.RaiseOthers, SendOptions.SendReliable);
             }
         }
     }