+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.1.0
+### Changes
+- Changed style of voting menu
+- Allow configuration to disable the voting in the lobby menu
+- Allow configuration to not play the same map twice in a row
+- Add colors for Stronghold and Hospital
+
+### Bugfixes
+- Fixed an issue when noone was voting in the lobby -> Defaults to randomly select any map
+- Fixed an issue where the voting amount would not be visible on maps with a long name
+
 # Version 1.0.1
 ### Changes
 - Changed "Time Left" Label to format "01:00 Seconds left"